Update: Pass selected text via Provider instead of props.
